split pycoverage tests and gcov tests into seperate files
